+++ R/cencorreg.R
@@ -3,7 +3,7 @@
 #' @description Computes three parametric correlation coefficients for one X variable and the corresponding R squared for multiple X variables, and a regression equation for censored data.
 #' @param y.var The column of y (response variable) values plus detection limits.
 #' @param cen.var The column of indicators, where 1 (or `TRUE`) indicates a detection limit in the `y.var` column, and 0 (or `FALSE`) indicates a detected value is in `y.var`.
-#' @param x.vars One or more uncensored explanatory variable(s). See Details
+#' @param x.vars One or more uncensored explanatory variable(s). For multiple variables it must be a data frame of numeric, character and factor variables. See Details
 #' @param LOG Indicator of whether to compute the regression in the original y units, or on their logarithms.  The default is to use the logarithms (`LOG = TRUE`).  To compute in original units, specify the option `LOG = FALSE` (or `LOG = 0`).
 #' @param verbose default `verbose=2`, see details.
 #' @export
@@ -57,14 +57,12 @@ cencorreg <- function(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
   #         verbose = 0 prints no output or plots.
 
   if (LOG == TRUE)  {lnvar <- log(nonas[,1])    # take logs of Y (default)
-  flip.log <- max(lnvar) +1 - lnvar
-  #  print(max(lnvar)+1)
-  surv.log <- Surv(flip.log, as.logical(nonas[,2]) )
+  flip.log <- -1 * lnvar
+  surv.log <- Surv(flip.log, as.logical(1-nonas[,2]) )
 
   if (is.data.frame(xnona))  {             # multiple x variables
     reg.out <- survreg(surv.log ~ ., data = xnona, dist = "gaussian")
     cn <- names(reg.out$coefficients[-1])
-    #    print(cn)
     xvars.txt <- cn[1]
     for (i in 1:length(cn))  {j <-(i+1)
     if (i != length(cn)) xvars.txt <- paste(xvars.txt, cn[j], sep = "+")
@@ -80,10 +78,9 @@ cencorreg <- function(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
   reg.out$call[3] <- xname
   }
 
-  ylog.pred <- max(lnvar) +1 - reg.out$linear.predictors
+  ylog.pred <- -1 * reg.out$linear.predictors
   reg.out$call[2] <- paste("log(", yname, ")", sep = "")
   reg.out$coefficients <- reg.out$coefficients * (-1)
-  reg.out$coefficients[1] <- max(lnvar)+1 + reg.out$coefficients[1]   # coeffs in cenreg lognormal
   ylog.resi <- lnvar - ylog.pred
   reg.out$linear.predictors <- ylog.pred
   reg.out$resids <- ylog.resi
@@ -120,7 +117,7 @@ cencorreg <- function(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
   reg.out$call[2] <- yname
   ynorm.pred <- reg.out$linear.predictors
   ynorm.resi <- y.low - ynorm.pred
-  reg.out$linear.predictors <- ynorm.pred
+  # reg.out$linear.predictors <- ynorm.pred
   reg.out$resids <- ynorm.resi
   vtext<- paste("Quantiles of", yname, "residuals")
 
@@ -132,7 +129,7 @@ cencorreg <- function(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
   } }  # end of low censored = 0
 
     #  negative y values.  Use flip variable to set -Inf as low end for censored values
-    else{ flip.norm <- max(nonas[,1]) +1 - nonas[,1]
+    else{ flip.norm <- -1 * nonas[,1]
     surv.norm <- Surv(flip.norm, as.logical(1-nonas[,2]) )
 
     if (is.data.frame(xnona))  {             # multiple x variables
@@ -154,13 +151,11 @@ cencorreg <- function(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
     }
 
     reg.out$call[2] <- yname
-    ynorm.pred <- max(nonas[,1]) +1 - reg.out$linear.predictors
+    ynorm.pred <- -1 * reg.out$linear.predictors
     reg.out$coefficients <- reg.out$coefficients * (-1)
-    reg.out$coefficients[1] <- max(nonas[,1]) +1 + reg.out$coefficients[1]   # coeffs in orig scale
     ynorm.resi <- nonas[,1] - ynorm.pred
     reg.out$linear.predictors <- ynorm.pred
     reg.out$resids <- ynorm.resi
-
     vtext<- paste("Quantiles of", yname, "residuals")
 
     if (verbose == 2) {
@@ -174,7 +169,7 @@ cencorreg <- function(y.var, cen.var, x.vars, LOG = TRUE, verbose = 2)
 
   if (is.data.frame(xnona))       {   # multiple x variables.  Print r-squared.
     LRr2 <- signif(1-exp(-2*(reg.out$loglik[2]-reg.out$loglik[1])/length(reg.out$y)),4)
-    McFr2 <- signif((1-reg.out$loglik[2]/reg.out$loglik[1]),4)
+    McFr2 <- 1-(reg.out$loglik[2]/reg.out$loglik[1])
     Nag.r2 <- signif((1-exp(-2*(reg.out$loglik[2]-reg.out$loglik[1])/length(reg.out$y))) /(1-exp(2*reg.out$loglik[1]/length(reg.out$y))),4)
     AIC <- -2*reg.out$loglik[2] + (2*reg.out$df +1)
     BIC <- -2*reg.out$loglik[2] + log(reg.out$df+reg.out$df.residual+1)*reg.out$df